third_party/blink/web_tests/fast/forms/select/listbox-tap-input-change-event-expected.txt

Tapping on listbox items should fire an input and a change event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS changeEventCounter is 0
PASS inputEventCounter is 0
PASS changeEventCounter is 1
PASS inputEventCounter is 1
PASS successfullyParsed is true

TEST COMPLETE

1
2
3
